pyproject.toml
[tool.poetry]
name = "strawberry-autograph"
version = "0.1.0"
description = "A strawberry graphql schema wrapper with operation autogeneration."
authors = ["Arthur Ryzhak <ryzhakar@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/ryzhakar/strawberry-autograph"
repository = "https://github.com/ryzhakar/strawberry-autograph"
keywords = ["strawberry-graphql", "graphql", "testing", "autogeneration"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Testing"
]

[tool.poetry.dependencies]
python = "^3.10"
strawberry-graphql = {version = "^0.209", optional = true}

[tool.poetry.group.dev.dependencies]
wemake-python-styleguide = "^0.18.0"
devtools = "^0.12.2"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
